Broke out TerminusSpecificProductTypes class and removed unnecessary namespaces
