+++ mzLib/Transcriptomics/ClassExtensions.cs
@@ -1,9 +1,5 @@
 ﻿using Omics.Modifications;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using Transcriptomics.Digestion;
 
 namespace Transcriptomics
