+++ mzLib/Transcriptomics/RNA.cs
@@ -1,10 +1,5 @@
 ﻿using Chemistry;
 using Omics.Modifications;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Transcriptomics
 {
